Move sampler to common location and rename to SamplerConfig
